warnings
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@8405 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/game/NetHandler.cxx
@@ -596,7 +596,7 @@ void NetHandler::udpSend(const void *b, size_t l) {
   } else {
     // Buffer new data
     memcpy(&udpOutputBuffer[udpOutputLen], (const char *)b, l);
-    udpOutputLen += l;
+    udpOutputLen += (int)l;
     // Send buffer if is almost full
     if (udpOutputLen > sizeLimit) {
       sendto(udpSocket, udpOutputBuffer, udpOutputLen, 0,
